+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ## Bug Fix
 - Clicking on a search result while on kana writing mode now searches for that result (before, it only worked while on romaji mode)
+- Fixed issue where list of highlighted kanji wouldn't show up if extension popup was opened too soon
 
 # [Changelog v0.6.2](https://github.com/digas99/wanikani-kanji-highlighter/releases/tag/v0.6.2)
 Released on 29/12/2021